requirements.txt
selenium
pytest
pytest-html
faker
python-dotenv
webdriver-manager